# docs/openapi.yaml
openapi: "3.1.0"
info:
  version: "1.0.0"
  title: "SkillSwap API"
  description: Marketplace interno de troca de mentorias
...
servers:
  - url: http://localhost:5000
    description: Servidor de Desenvolvimento Local
  - url: https://api.skillswap.com
    description: Servidor de Produção
paths:
...
  /api/v1/usuario/{id}:
    parameters:
      - name: id
        description: A identificação única do usuário
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UsuarioId"
    get:
      tags:
        - Usuários (Admin)
      summary: Lê um usuário
      responses:
        "200":
          description: O usuário correspondente ao `UsuarioId` providenciado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsuarioDTO"
        "404":
          description: Nenhum usuário encontrado com o `UsuarioId` providenciado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Erro inesperado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      tags:
        - Usuários (Admin)
      summary: Atualiza dados básicos do usuário
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UsuarioId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AtualizarUsuarioRequestDTO"
      responses:
        "200":
          description: Usuário atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsuarioDTO"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags:
        - Usuários (Admin)
      summary: Remove um colaborador
      description: De acordo com a regra, a remoção exige consentimento ou ação do ADM
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UsuarioId"
      responses:
        "204":
          description: Usuário removido com sucesso (sem conteúdo)
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/usuario/{id}/verificacao/enviar:
    post:
      tags:
        - Usuários (Colaborador)
      summary: Envia o código de verificação de duas etapas
      description: Gera um código de 6 dígitos e envia para o email ou SMS do usuário.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UsuarioId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnviarCodigoRequestDTO"
      responses:
        "200":
          description: Código enviado com sucesso
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          description: Método de verificação inválido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/usuario/{id}/verificacao/validar:
    post:
      tags:
        - Usuários (Colaborador)
      summary: Valida o código de verificação informado
      description: Verifica se o código é válido e não expirou.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UsuarioId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidarCodigoRequestDTO"
      responses:
        "200":
          description: Código validado com sucesso
        "400":
          description: Código inválido ou expirado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/usuario:
    post:
      tags:
        - Usuários (Admin)
      summary: Cadastra um novo usuário
      description: Cria uma nova conta e retorna os dados do usuário gerado.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CriarUsuarioRequestDTO"
      responses:
        "201":
          description: Usuário criado com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsuarioDTO"
        "400":
          description: Erro de validação
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Erro inesperado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/v1/usuario/{id}/informacoes:
    get:
      tags:
        - Usuários (Admin)
      summary: Busca informações detalhadas (contato e endereço)
      description: Dados provenientes da tabela INFORMACOES do DER
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UsuarioId"
      responses:
        "200":
          description: Retorna informações detalhadas do usuário
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InformacoesDTO"

    put:
      tags:
        - Usuários (Admin)
      summary: Atualiza informações de contato e endereço
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UsuarioId"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InformacoesDTO"
      responses:
        "200":
          description: Informações atualizadas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InformacoesDTO"
...
components:
  schemas:
...
    UsuarioId:
      description: O identificador único do usuário (UUID)
      type: string
      format: uuid
    UsuarioDTO:
      type: object
      required:
        - id
        - nome
        - role
      properties:
        id:
          $ref: "#/components/schemas/UsuarioId"
        nome:
          type: string
        role:
          type: string
          enum:
            - admin
            - usuario
...
    CriarUsuarioRequestDTO:
      type: object
      required:
        - nome
        - email
        - telefone
        - senha
      properties:
        nome:
          type: string
          description: Nome completo do usuário
        email:
          type: string
          format: email
          description: Email de acesso do usuário
        telefone:
          type: string
          description: Telefone de contato (apenas números)
        senha:
          type: string
          format: password
    AtualizarUsuarioRequestDTO:
      type: object
      properties:
        nome:
          type: string
        role:
          type: string
          enum: [admin, usuario]
    EnviarCodigoRequestDTO:
      type: object
      required:
        - metodo
      properties:
        metodo:
          type: string
          enum: [email]
          description: Canal de envio do código de verificação
    ValidarCodigoRequestDTO:
      type: object
      required:
        - codigo
      properties:
        codigo:
          type: string
          minLength: 6
          maxLength: 6
          description: Código de 6 dígitos recebido pelo usuário
...
    InformacoesDTO:
      type: object
      description: Mapeamento completo da tabela INFORMACOES
      properties:
        email:
          type: string
          format: email
        telefone:
          type: string
        endereco_rua:
          type: string
        endereco_numero:
          type: string
        endereco_bairro:
          type: string
        endereco_cidade:
          type: string
        endereco_estado:
          type: string
        endereco_cep:
          type: string
...
    Error:
      type: object
      required:
        - mensagem
      properties:
        mensagem:
          description: Uma mensagem de erro compreensível para o usuário
          type: string
  responses:
    NotFound:
      description: O recurso solicitado não foi encontrado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: X-Api-Key
security:
  - ApiKey: []